--- frontend/my-app/src/Component/any.module.css
@@ -1,162 +1,0 @@
-.product {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    grid-gap: 1rem;
-    padding: 1rem;
-    box-sizing: border-box;
-}
-
-.imageContainer {
-    position: relative;
-    display: inline-block;
-}
-
-.image {
-    width: 30rem;
-    height: 30rem;
-    background-size: cover;
-    margin-left: 20%;
-    border-radius: 1rem;
-    position: absolute;
-}
-
-.imageButtons .next {
-    position: absolute;
-    top: 150%;
-    right: 8%;
-    background-image: url(../Images/but-right.png);
-    border: none;
-    outline: none;
-    cursor: pointer;
-    padding: 1.5rem;
-    border-radius: 2rem;
-    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.5);
-}
-
-.imageButtons .pre {
-    position: absolute;
-    top: 150%;
-    right: 75%;
-    background-image: url(../Images/but-left.png);
-    border: none;
-    outline: none;
-    cursor: pointer;
-    padding: 1.5rem;
-    border-radius: 2rem;
-    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.5);
-}
-
-.description {
-    width: 100%;
-    height: 100%;
-    background-color: #fff;
-    padding: 1rem;
-    padding-right: 20%;
-    box-sizing: border-box;
-}
-
-.description h1 {
-    font-size: 2rem;
-    font-weight: bold;
-    color: #cf1111;
-    margin-bottom: 1rem;
-    text-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.5);
-}
-
-.price {
-    span {
-        font-size: 1.5rem;
-        font-weight: bold;
-        color: #cf1111;
-        text-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.5);
-    }
-}
-
-.require button {
-    cursor: pointer;
-    padding: 1rem 1rem;
-    border-radius: 2rem;
-    font-size: 1.3rem;
-    font-weight: bold;
-}
-
-.require button:hover {
-    background-color: rgb(56, 140, 76);
-    color: #fff;
-    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.5);
-}
-
-.chatPopup {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.closeButton {
-    position: absolute;
-    top: 1rem;
-    right: 1rem;
-    width: 2rem;
-    height: 2rem;
-    background-image: url(../Images/close-but.png);
-    background-repeat: no-repeat;
-    background-size: contain;
-    cursor: pointer;
-}
-
-.chat {
-    width: 30rem;
-    height: 30rem;
-    background-color: #fff;
-    overflow: auto;
-    padding: 1rem;
-    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.4);
-}
-
-.chatMessage {
-    margin-bottom: 1rem;
-}
-
-.chatMessageContent {
-    background-color: #f7f7f7;
-    padding: 1rem;
-    border-radius: 1rem;
-    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.4);
-}
-
-.send {
-    align-self: flex-end;
-}
-
-.fieldSend {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-top: 1rem;
-}
-
-.fieldSend input {
-    flex: 1;
-    padding: 1rem;
-    border-radius: 2rem;
-    border: none;
-    outline: none;
-    font-size: 1.3rem;
-}
-
-.fieldSend button {
-    width: 3rem;
-    height: 3rem;
-    background-image: url(../Images/send.png);
-    background-repeat: no-repeat;
-    background-size: contain;
-    border: none;
-    outline: none;
-    cursor: pointer;
-}
