--- frontend/my-app/src/Component/categories.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Categories = () => {
-    return (
-        <div>
-            <h1>Categories</h1>
-        </div>
-)
-}
-
-export default Categories
