+++ frontend/my-app/src/Pages/Chat/chat.module.css
@@ -77,7 +77,7 @@
 .sendButton {
     width: 3rem;
     height: 3rem;
-    background-image: url(../Images/send.png);
+    background-image: url(../../Images/send.png);
     background-repeat: no-repeat;
     background-size: contain;
     border: none;
