+++ frontend/my-app/src/Pages/Personal/personal.module.css
@@ -31,10 +31,43 @@
 
 .contentname,
 .information {
-    top: 50px;
+    top: 30px;
     position: absolute
 }
 
+.information input {
+    border-radius: var(--br-31xl);
+    background-color: var(--color-darkslateblue);
+    color: var(--color-sandybrown);
+    width: 200px;
+    height: 40px;
+    mix-blend-mode: difference;
+    font-size: 20px;
+    border: 0;
+    z-index: 2
+}
+
+.information button {
+    height: 67px;
+    width: 50%;
+    border-radius: 50px;
+    padding: 10px 30px;
+    background-color: var(--color-darkslateblue);
+    color: var(--color-whitesmoke);
+    display: block;
+    text-align: center;
+    cursor: pointer;
+    font-size: 20px;
+    margin: 10px;
+    border: 0;
+    z-index: 5
+}
+
+.information button:hover {
+    color: var(--color-darkslateblue);
+    background-color: var(--color-whitesmoke);
+    font-size: 30px;
+}
 .contentname {
     display: flex;
     -webkit-box-orient: vertical;
@@ -69,12 +102,12 @@
     font-weight: 600;
     margin-bottom: 150px;
     text-align: left;
-    width: 400px
+    width: 500px
 }
 
 .information {
     left: 130px;
-    height: 600px;
+    height: 630px;
     border-radius: var(--br-31xl);
     width: 500px;
     background-color: var(--color-sandybrown);
@@ -104,7 +137,7 @@
 .address,
 .mail,
 .phone,
-.website {
+.name {
     margin: 50px
 }
